test: Rework how the OCaml listening port is passed to SC
Instead of passing the OCaml listening port as the first argument in the
OSC packet, write it as raw bytes to a file at test.data, and then read
the file in SuperCollider.

This enables us to test specific messages, rather than having to add an
argument to each test message.

Signed-off-by: John Else <john.else@citrix.com>